refactor prove
